Implement Route() exchange runner
Route is an exchange runner that uses consistent hashing algorithms to
spread incoming data sets between the nodes consistently. It uses the
first column of an incoming data set to determine the target of this
row, and this target will remain consistent until the cluster is
changed. Even when cluster changes, the same rows most likely continue
going to the same nodes.

The first column of an incoming data set is discarded; it is used only
to determine the target node for every row.

Data sent back from target nodes can be gathered, but the order does not
have to remain the same.